+++ examples/Start.test.tsx
@@ -22,13 +22,13 @@ beforeEach(async () => {
 
 test("should display the next month button", async () => {
   expect(nextButton()).toBeVisible();
-  expect(nextButton()).not.toHaveAttribute("disabled");
+  expect(nextButton()).not.toHaveAttribute("aria-disabled", "true");
   expect(nextButton()).not.toHaveAttribute("disabled", "true");
 });
 
 test("should display the previous month buttons", async () => {
   expect(previousButton()).toBeVisible();
-  expect(previousButton()).not.toHaveAttribute("disabled");
+  expect(previousButton()).not.toHaveAttribute("aria-disabled", "true");
   expect(previousButton()).not.toHaveAttribute("disabled", "true");
 });