+++ src/components/Root.tsx
@@ -1,4 +1,4 @@
-import React, { type HTMLAttributes } from "react";
+import React, { type Ref, type HTMLAttributes } from "react";
 
 /**
  * Render the root element.
@@ -6,8 +6,14 @@ import React, { type HTMLAttributes } from "react";
  * @group Components
  * @see https://daypicker.dev/guides/custom-components
  */
-export function Root(props: HTMLAttributes<HTMLDivElement>) {
-  return <div {...props} />;
+export function Root(
+  props: {
+    /** Ref for root element used when `animate` is `true`. */
+    rootRef?: Ref<HTMLDivElement>;
+  } & HTMLAttributes<HTMLDivElement>
+) {
+  const { rootRef, ...rest } = props;
+  return <div {...rest} ref={rootRef} />;
 }
 
 export type RootProps = Parameters<typeof Root>[0];