+++ src/helpers/calculateFocusTarget.ts
@@ -2,6 +2,21 @@ import { DayFlag } from "../UI.js";
 import type { CalendarDay } from "../classes/index.js";
 import type { Modifiers } from "../types/index.js";
 
+enum FocusTargetPriority {
+  Today = 0,
+  Selected,
+  LastFocused,
+  FocusedModifier
+}
+
+function isFocusableDay(modifiers: Modifiers) {
+  return (
+    !modifiers[DayFlag.disabled] &&
+    !modifiers[DayFlag.hidden] &&
+    !modifiers[DayFlag.outside]
+  );
+}
+
 export function calculateFocusTarget(
   days: CalendarDay[],
   getModifiers: (day: CalendarDay) => Modifiers,
@@ -10,42 +25,42 @@ export function calculateFocusTarget(
 ) {
   let focusTarget: CalendarDay | undefined;
 
-  let index = 0;
-  let found = false;
-
-  while (index < days.length && !found) {
-    const day = days[index];
+  let foundFocusTargetPriority: FocusTargetPriority | -1 = -1;
+  for (const day of days) {
     const modifiers = getModifiers(day);
 
-    if (
-      !modifiers[DayFlag.disabled] &&
-      !modifiers[DayFlag.hidden] &&
-      !modifiers[DayFlag.outside]
-    ) {
-      if (modifiers[DayFlag.focused]) {
+    if (isFocusableDay(modifiers)) {
+      if (
+        modifiers[DayFlag.focused] &&
+        foundFocusTargetPriority < FocusTargetPriority.FocusedModifier
+      ) {
         focusTarget = day;
-        found = true;
-      } else if (lastFocused?.isEqualTo(day)) {
+        foundFocusTargetPriority = FocusTargetPriority.FocusedModifier;
+      } else if (
+        lastFocused?.isEqualTo(day) &&
+        foundFocusTargetPriority < FocusTargetPriority.LastFocused
+      ) {
         focusTarget = day;
-        found = true;
-      } else if (isSelected(day.date)) {
+        foundFocusTargetPriority = FocusTargetPriority.LastFocused;
+      } else if (
+        isSelected(day.date) &&
+        foundFocusTargetPriority < FocusTargetPriority.Selected
+      ) {
         focusTarget = day;
-        found = true;
-      } else if (modifiers[DayFlag.today]) {
+        foundFocusTargetPriority = FocusTargetPriority.Selected;
+      } else if (
+        modifiers[DayFlag.today] &&
+        foundFocusTargetPriority < FocusTargetPriority.Today
+      ) {
         focusTarget = day;
-        found = true;
+        foundFocusTargetPriority = FocusTargetPriority.Today;
       }
     }
-
-    index++;
   }
 
   if (!focusTarget) {
     // return the first day that is focusable
-    focusTarget = days.find((day) => {
-      const m = getModifiers(day);
-      return !m[DayFlag.disabled] && !m[DayFlag.hidden] && !m[DayFlag.outside];
-    });
+    focusTarget = days.find((day) => isFocusableDay(getModifiers(day)));
   }
   return focusTarget;
 }